requirements.txt
streamlit>=1.36
ViennaRNA>=2.6 
